# web/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Artemis Studio API",
    "description" : "Cluster-wide management and observability for Apache ActiveMQ Artemis",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/clusters/{clusterId}/queues" : {
      "get" : {
        "tags" : [ "cluster-resource-controller" ],
        "operationId" : "queues",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ResourceQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedViewQueueView"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "queue-lifecycle-controller" ],
        "operationId" : "createQueue",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateQueueRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/diverts" : {
      "get" : {
        "tags" : [ "routing-controller" ],
        "operationId" : "diverts",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ResourceQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedViewDivertView"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "routing-controller" ],
        "operationId" : "createDivert",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateDivertRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DivertMutationView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/addresses" : {
      "get" : {
        "tags" : [ "cluster-resource-controller" ],
        "operationId" : "addresses",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ResourceQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedViewAddressView"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "queue-lifecycle-controller" ],
        "operationId" : "createAddress",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAddressRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}" : {
      "get" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClusterDetail"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "delete_4",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/v1/clusters/{clusterId}/topology" : {
      "get" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "topology",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TopologyView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/sessions" : {
      "get" : {
        "tags" : [ "cluster-resource-controller" ],
        "operationId" : "sessions",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ResourceQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedViewSessionView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/producers" : {
      "get" : {
        "tags" : [ "cluster-resource-controller" ],
        "operationId" : "producers",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ResourceQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedViewProducerView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/consumers" : {
      "get" : {
        "tags" : [ "cluster-resource-controller" ],
        "operationId" : "consumers",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ResourceQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedViewConsumerView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clusters/{clusterId}/connections" : {
      "get" : {
        "tags" : [ "cluster-resource-controller" ],
        "operationId" : "connections",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ResourceQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedViewConnectionView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/bridges" : {
      "get" : {
        "tags" : [ "routing-controller" ],
        "operationId" : "bridges",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ResourceQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedViewBridgeView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/diverts/{name}" : {
      "delete" : {
        "tags" : [ "routing-controller" ],
        "operationId" : "deleteDivert",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TopologyView" : {
        "type" : "object",
        "properties" : {
          "clusterId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "nodes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LogicalNodeView"
            }
          }
        },
        "required" : [ "clusterId", "nodes" ]
      },
      "ClusterDetail" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "topology" : {
            "$ref" : "#/components/schemas/TopologyView"
          },
          "capabilities" : {
            "$ref" : "#/components/schemas/CapabilitiesView"
          },
          "health" : {
            "$ref" : "#/components/schemas/HealthView"
          },
          "environmentId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          }
        },
        "required" : [ "capabilities", "health", "id", "name", "topology" ]
      },
...
      "CreateQueueRequest" : {
        "type" : "object",
        "properties" : {
          "address" : {
            "type" : "string",
            "description" : "The address this queue binds to.",
            "minLength" : 1
          },
          "name" : {
            "type" : "string",
            "description" : "The queue's name.",
            "minLength" : 1
          },
          "routingType" : {
            "type" : "string",
            "description" : "ANYCAST or MULTICAST. Cannot be changed once the queue exists.",
            "enum" : [ "ANYCAST", "MULTICAST" ],
            "minLength" : 1,
            "pattern" : "(?i)ANYCAST|MULTICAST"
          },
          "durable" : {
            "type" : "boolean",
            "default" : true,
            "description" : "Whether the queue survives a broker restart."
          },
          "filter" : {
            "type" : [ "string", "null" ],
            "description" : "A JMS selector limiting what the queue accepts."
          },
          "maxConsumers" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "description" : "Maximum concurrent consumers; -1 for unlimited."
          },
          "purgeOnNoConsumers" : {
            "type" : [ "boolean", "null" ],
            "description" : "Delete the queue when its last consumer disconnects."
          },
          "exclusive" : {
            "type" : [ "boolean", "null" ],
            "description" : "Route to exactly one consumer at a time."
          },
          "nonDestructive" : {
            "type" : [ "boolean", "null" ],
            "description" : "Messages stay on the queue after delivery."
          },
          "ringSize" : {
            "type" : [ "integer", "null" ],
            "format" : "int64",
            "description" : "Retain at most this many messages; -1 for unlimited."
          },
          "autoCreateAddress" : {
            "type" : "boolean",
            "default" : true,
            "description" : "Create the address if it does not exist."
          }
        },
        "required" : [ "address", "name", "routingType" ]
      },
      "LifecycleOutcomeView" : {
        "type" : "object",
        "properties" : {
          "dryRun" : {
            "type" : "boolean",
            "description" : "Whether this was a preview that mutated nothing."
          },
          "cap" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "The bulk safety cap in force."
          },
          "overCap" : {
            "type" : "boolean",
            "description" : "Whether the summed estimate exceeds the cap."
          },
          "partial" : {
            "type" : "boolean",
            "description" : "Whether some nodes applied the command and others did not."
          },
          "totalAffected" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Total messages destroyed, or that would be."
          },
          "nodes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NodeOutcomeView"
            }
          }
        },
        "required" : [ "cap", "dryRun", "nodes", "overCap", "partial", "totalAffected" ]
      },
...
      "CreateDivertRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "The divert's unique name on each node.",
            "minLength" : 1
          },
          "routingName" : {
            "type" : [ "string", "null" ],
            "description" : "The routing name; defaults to the divert's name."
          },
          "address" : {
            "type" : "string",
            "description" : "The address whose messages are diverted.",
            "minLength" : 1
          },
          "forwardingAddress" : {
            "type" : "string",
            "description" : "The address messages are diverted to.",
            "minLength" : 1
          },
          "exclusive" : {
            "type" : "boolean",
            "default" : false,
            "description" : "Take the message rather than copy it. Exclusive diverts are evaluated before non-exclusive ones."
          },
          "filter" : {
            "type" : [ "string", "null" ],
            "description" : "A filter limiting which messages are diverted."
          },
          "routingType" : {
            "type" : [ "string", "null" ],
            "description" : "Routing type applied to the diverted copy.",
            "enum" : [ "ANYCAST", "MULTICAST", "PASS", "STRIP" ]
          }
        },
        "required" : [ "address", "forwardingAddress", "name" ]
      },
      "DivertMutationView" : {
        "type" : "object",
        "properties" : {
          "outcome" : {
            "$ref" : "#/components/schemas/LifecycleOutcomeView"
          },
          "brokerXml" : {
            "type" : "string",
            "description" : "The <divert> element that would make this broker's own configuration carry the divert. A divert created over management persists across restarts but is absent from configuration, and this is what closes that gap."
          }
        },
        "required" : [ "brokerXml", "outcome" ]
      },
...
      "CreateAddressRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "The address name.",
            "minLength" : 1
          },
          "routingTypes" : {
            "type" : "string",
            "description" : "Comma-separated routing types the address supports.",
            "enum" : [ "ANYCAST", "MULTICAST", "ANYCAST,MULTICAST" ],
            "minLength" : 1,
            "pattern" : "(?i)ANYCAST|MULTICAST|ANYCAST,MULTICAST|MULTICAST,ANYCAST"
          }
        },
        "required" : [ "name", "routingTypes" ]
      },
...
      "ResourceQuery" : {
        "type" : "object",
        "properties" : {
          "q" : {
            "type" : "string"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "type" : "string"
          }
        }
      },
      "PagedViewSessionView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SessionView"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "count", "data", "page", "pageSize" ]
      },
...
      "PagedViewQueueView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QueueView"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "count", "data", "page", "pageSize" ]
      },
...
      "PagedViewProducerView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProducerView"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "count", "data", "page", "pageSize" ]
      },
...
      "PagedViewDivertView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DivertView"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "count", "data", "page", "pageSize" ]
      },
...
      "PagedViewConsumerView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConsumerView"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "count", "data", "page", "pageSize" ]
      },
...
      "PagedViewConnectionView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConnectionView"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "count", "data", "page", "pageSize" ]
      },
...
      "PagedViewBridgeView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BridgeView"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "count", "data", "page", "pageSize" ]
      },
...
      "AddressView" : {
        "type" : "object",
        "properties" : {
          "nodeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "nodeName" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "routingTypes" : {
            "type" : [ "string", "null" ]
          },
          "queueCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "messageCount" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "messageCount", "name", "nodeId", "nodeName", "queueCount" ]
      },
      "PagedViewAddressView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AddressView"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "count", "data", "page", "pageSize" ]
      },
...
    }
  }
}